move simulations case to examples directory
