+++ examples/simulation creator_multienergyCase.py
